requirements.txt
torch==2.1.0
torchvision==0.16.0
medmnist==3.0.1
opencv-python-headless==4.8.1.78
streamlit==1.28.0
umap-learn==0.5.3
scikit-learn==1.3.2
matplotlib==3.8.2
numpy==1.26.2
scipy==1.11.4
tqdm==4.66.1
Pillow==10.1.0